# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auction API",
        "description": "API for the Antique Items Auction Web Application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bid-configs": {
            "get": {
                "tags": [
                    "Bid Configurations"
                ],
                "summary": "Get all bid configurations",
                "description": "Retrieves a list of all bid configurations",
                "operationId": "get_bid_config_index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BidConfig"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bid Configurations"
                ],
                "summary": "Create a new bid configuration",
                "description": "Creates a new bid configuration with the provided data",
                "operationId": "post_bid_config_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BidConfigType"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bid configuration created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BidConfig"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/bid-configs/me": {
            "get": {
                "tags": [
                    "Bid Configurations"
                ],
                "summary": "Get current user's bid configuration",
                "description": "Retrieves the bid configuration for the authenticated user",
                "operationId": "get_bid_config_me",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BidConfig"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/bid-configs/{uuid}": {
            "get": {
                "tags": [
                    "Bid Configurations"
                ],
                "summary": "Get a specific bid configuration",
                "description": "Retrieves details of a specific bid configuration by UUID",
                "operationId": "get_bid_config_show",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the bid configuration",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BidConfig"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bid configuration not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bid Configurations"
                ],
                "summary": "Update an existing bid configuration",
                "description": "Updates an existing bid configuration with the provided data",
                "operationId": "put_bid_config_update",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the bid configuration to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BidConfigType"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bid configuration updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BidConfig"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bid configuration not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bid Configurations"
                ],
                "summary": "Delete a bid configuration",
                "description": "Deletes a bid configuration by UUID",
                "operationId": "delete_bid_config_delete",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the bid configuration to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bid configuration deleted successfully"
                    },
                    "404": {
                        "description": "Bid configuration not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/bids": {
            "get": {
                "tags": [
                    "Bids"
                ],
                "summary": "Get a list of bids",
                "description": "Retrieves a paginated list of all bids",
                "operationId": "get_bid_index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Bid"
                                            }
                                        },
                                        "totalItems": {
                                            "type": "integer"
                                        },
                                        "itemsPerPage": {
                                            "type": "integer"
                                        },
                                        "totalPages": {
                                            "type": "integer"
                                        },
                                        "currentPage": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bids"
                ],
                "summary": "Create a new bid",
                "description": "Creates a new bid with the provided data",
                "operationId": "post_bid_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BidType"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bid created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bid2"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/bids/{uuid}": {
            "get": {
                "tags": [
                    "Bids"
                ],
                "summary": "Get a specific bid",
                "description": "Retrieves details of a specific bid by UUID",
                "operationId": "get_bid_show",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the bid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bid2"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bid not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bids"
                ],
                "summary": "Update an existing bid",
                "description": "Updates an existing bid with the provided data",
                "operationId": "put_bid_update",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the bid to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BidType"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bid updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bid2"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bid not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bids"
                ],
                "summary": "Delete a bid",
                "description": "Deletes a bid by UUID",
                "operationId": "delete_bid_delete",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the bid to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bid deleted successfully"
                    },
                    "404": {
                        "description": "Bid not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BidConfig": {
                "required": [
                    "user",
                    "maxBidAmount",
                    "id",
                    "uuid",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/User3"
                    },
                    "maxBidAmount": {
                        "type": "string"
                    },
                    "bidAlertPercentage": {
                        "type": "integer",
                        "nullable": true
                    },
                    "reservedAmount": {
                        "type": "string",
                        "nullable": true
                    },
                    "id": {
                        "type": "integer"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            },
            "BidConfigType": {
                "required": [
                    "user",
                    "maxBidAmount"
                ],
                "properties": {
                    "user": {
                        "type": "string"
                    },
                    "maxBidAmount": {
                        "type": "number"
                    },
                    "bidAlertPercentage": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Bid": {
                "required": [
                    "bidder",
                    "item",
                    "isAutoBid",
                    "amount",
                    "id",
                    "uuid",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "bidder": {
                        "$ref": "#/components/schemas/User4"
                    },
                    "item": {
                        "$ref": "#/components/schemas/Item4"
                    },
                    "bidTime": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "isAutoBid": {
                        "type": "boolean"
                    },
                    "amount": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [],
...
}